+++ ai4rag/rag/foundation_models/foundation_model.py
@@ -2,9 +2,7 @@
 # Copyright IBM Corp. 2025
 # SPDX-License-Identifier: Apache-2.0
 # -----------------------------------------------------------------------------
-from abc import ABC, abstractmethod
-from typing import Any, Literal, Annotated
-from string import Formatter
+from typing import Any, Annotated
 
 from annotated_types import Gt, Le, Ge
 from pydantic import BaseModel
@@ -13,70 +11,13 @@ from llama_stack_client import LlamaStackClient
 from llama_stack_client.types import SystemMessage, UserMessage
 
 from ai4rag.search_space.src.model_props import (
-    CONTEXT_TEXT_PLACEHOLDER,
-    QUESTION_PLACEHOLDER,
-    REFERENCE_DOCUMENTS_PLACEHOLDER,
     get_system_message_text,
     get_user_message_text,
-    get_context_template_text,
 )
 from ai4rag.utils.constants import ChatGenerationConstants
 
-
-def _validate_prompt_templates_placeholders(
-        template_str: str,
-        template_name: Literal["context_template_text", "user_message_text"],
-) -> str:
-    """
-    Validates if user provided correct placeholders in given template text in respect to default placeholders.
-
-    Parameters
-    ----------
-    template_str : str
-        Prompt template with proper placeholders to be validated.
-
-    template_name : Literal["context_template_text", "user_message_text"]
-        Name of the template that will be validated. Used for required placeholders selection.
-
-    Returns
-    -------
-    str
-        Prompt template with filled placeholders.
-
-    Raises
-    ------
-    ValueError
-        When user provided less placeholders than expected.
-
-        When user provided wrong placeholder name.
-    """
-    if template_name == "context_template_text":
-        required_placeholders = (CONTEXT_TEXT_PLACEHOLDER,)
-    elif template_name == "user_message_text":
-        required_placeholders = (QUESTION_PLACEHOLDER, REFERENCE_DOCUMENTS_PLACEHOLDER)
-    else:
-        raise ValueError(f"Cannot validate presence of expected template placeholders on field: {template_name}")
-
-    placeholders_count = 0
-
-    for _, field_name, _, _ in Formatter().parse(template_str):
-        if field_name is None:
-            # when there is text NOT followed by a placeholder template
-            continue
-        if field_name not in required_placeholders:
-            raise ValueError(
-                f"Custom {field_name.split('_')[0]} template text got unexpected placeholder `{field_name}`, "
-                f"valid placeholders are `{required_placeholders}`."
-            )
-
-        placeholders_count += 1
-
-    if placeholders_count != len(required_placeholders):
-        raise ValueError(
-            f"Incorrect number of placeholders required for {template_name.split('_')[0]} template text, "
-            f"expected {len(required_placeholders)} but got {placeholders_count}."
-        )
-    return template_str
+from .base import FoundationModel
+from .utils import _validate_prompt_templates_placeholders
 
 
 class ModelParameters(BaseModel):
@@ -84,39 +25,15 @@ class ModelParameters(BaseModel):
     temperature: Annotated[float, Ge(0), Le(1)] = ChatGenerationConstants.TEMPERATURE
 
 
-class FoundationModel(ABC):
-    def __init__(self, model_id: str, model_params: dict[str, Any]):
-        self.model_id = model_id
-        self.model_params = model_params
-
-    def __repr__(self) -> str:
-        return self.model_id
-
-    def __str__(self) -> str:
-        return repr(self)
-
-    def __eq__(self, other: Any) -> bool:
-        if not isinstance(other, FoundationModel):
-            return NotImplemented
-
-        return self.model_id == other.model_id
-
-    def __hash__(self):
-        return hash(self.model_id)
-
-    @abstractmethod
-    def chat(self, system_message: str, user_message: str) -> str:
-        """Docstring here"""
-
-
 class LlamaStackFoundationModel(FoundationModel):
     """Integration point to use any model via Llama-stack API / client"""
+
     def __init__(
-            self,
-            model_id: str,
-            model_params: dict[str, Any] | ModelParameters | None,
-            ls_client: LlamaStackClient,
-            **kwargs
+        self,
+        model_id: str,
+        model_params: dict[str, Any] | ModelParameters | None,
+        ls_client: LlamaStackClient,
+        **kwargs,
     ):
         super().__init__(model_id, model_params)
         self._ls_client = ls_client
@@ -172,7 +89,6 @@ class LlamaStackFoundationModel(FoundationModel):
         else:
             self._user_message_text = _validate_prompt_templates_placeholders(val, "user_message_text")
 
-
     def chat(self, system_message: str, user_message: str) -> str:
         """
         Chat completion for communication with selected foundation model.
@@ -195,7 +111,7 @@ class LlamaStackFoundationModel(FoundationModel):
             messages=[
                 SystemMessage(role="system", content=system_message),
                 UserMessage(role="user", content=user_message),
-            ]
+            ],
         )
         answer = response_chat.choices[0].message.content
